feat: merge Overview Status card into the Slots timeline
The Overview "Status" card largely duplicated the "Slots" timeline:
both showed the current slot, time-until-leader (with progress), and the
next leader slot. The only unique datum was the vote status.

Remove the Status card and surface the vote status as a compact badge in
the Slots timeline header, giving a more compact Overview without losing
information. The card grid drops from four to three columns at the widest
breakpoint accordingly.
