--- src/features/Overview/StatusCard/statusCard.module.css
@@ -1,11 +1,0 @@
-.stat-row {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--space-2);
-  width: 100%;
-
-  > div {
-    flex: 1;
-    min-width: 180px;
-  }
-}
